requirements.txt
flask==2.0.3
Werkzeug==2.0.3
jinja2==3.0.3
WTForms==2.3.3
flask_wtf==1.0.1

flask_migrate==3.1.0
flask_sqlalchemy==2.5.1
sqlalchemy==1.4.49

flask_login==0.5.0
celery==5.1.2
redis==4.5.5
colorama==0.4.6
PyJWT==2.4.0
WTForms-Alchemy==0.18.0

email_validator==1.3.1
blinker==1.5
python-dotenv==0.21.1

gunicorn==20.1.0
Flask-Minify==0.42
Flask-CDN==1.5.2

mysqlclient==2.1.1
